refactor: Move utility classes.
